Moved command line config into its own file
